+++ kind-integer/default.nix
@@ -1,9 +1,10 @@
 { mkDerivation, base, lib, singletons }:
 mkDerivation {
   pname = "kind-integer";
-  version = "0.5";
+  version = "0.6";
   src = lib.sources.cleanSource ./.;
   libraryHaskellDepends = [ base singletons ];
+  testHaskellDepends = [ base singletons ];
   homepage = "https://github.com/k0001/hs-kind";
   description = "Type-level integers. Like KnownNat, but for integers.";
   license = lib.licenses.bsd3;